+++ app/javascript/src/javascripts/autocomplete.js.erb
@@ -2,9 +2,9 @@ import Utility from './utility'
 
 let Autocomplete = {};
 
-Autocomplete.METATAGS = <%= Tag::METATAGS.to_json.html_safe %>;
+Autocomplete.METATAGS = <%= TagQuery::METATAGS.to_json.html_safe %>;
 Autocomplete.TAG_CATEGORIES = <%= TagCategory::MAPPING.to_json.html_safe %>;
-Autocomplete.ORDER_METATAGS = <%= Tag::ORDER_METATAGS.to_json.html_safe %>;
+Autocomplete.ORDER_METATAGS = <%= TagQuery::ORDER_METATAGS.to_json.html_safe %>;
 
 Autocomplete.TAG_PREFIXES = "-|~|" + Object.keys(Autocomplete.TAG_CATEGORIES).map(category => category + ":").join("|");
 Autocomplete.TAG_PREFIXES_REGEX = new RegExp("^(" + Autocomplete.TAG_PREFIXES + ")(.*)$", "i");
