+++ app/logical/tag_name_validator.rb
@@ -32,7 +32,7 @@ class TagNameValidator < ActiveModel::EachValidator
       record.errors.add(attribute, "'#{value}' cannot contain non-printable characters")
     when /\A[-~+_`(){}\[\]\/]/
       record.errors.add(attribute, "'#{value}' cannot begin with a '#{value[0]}'")
-    when /\A(#{Tag::METATAGS.join("|")}):(.+)\z/i
+    when /\A(#{TagQuery::METATAGS.join("|")}):(.+)\z/i
       record.errors.add(attribute, "'#{value}' cannot begin with '#{$1}:'")
     when /\A(#{Tag.categories.regexp}):(.+)\z/i
       record.errors.add(attribute, "'#{value}' cannot begin with '#{$1}:'")
