+++ app/presenters/post_presenter.rb
@@ -257,7 +257,7 @@ class PostPresenter < Presenter
   end
 
   def has_sequential_navigation?(params)
-    return false if Tag.has_metatag?(params[:q], :order)
+    return false if TagQuery.has_metatag?(params[:q], :order)
     return false if params[:pool_id].present? || params[:post_set_id].present?
     true
   end
